refactor(app): move components to each component folder (#20315)
* refactor(app): move components to each component folder
