+++ app/src/organisms/Desktop/ProtocolVisualization/LabwareSlotContainer/labwareslotcontainer.module.css
