+++ app/src/organisms/Desktop/ProtocolVisualization/StepDetailContainer/stepdetailcontainer.module.css
