+++ app/src/organisms/Desktop/ProtocolVisualization/TipPickupContainer/__tests__/TipPickupContainer.test.tsx
@@ -7,7 +7,7 @@ import { CLEAN, EMPTY } from '@opentrons/step-generation'
 import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 
-import { TipPickupContainer } from '../TipPickupContainer'
+import { TipPickupContainer } from '..'
 
 import type { ComponentProps } from 'react'
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
