+++ app/src/organisms/Desktop/ProtocolVisualization/TipPickupContainer/index.tsx
@@ -13,8 +13,8 @@ import {
 import { getLabwareViewBox } from '@opentrons/shared-data'
 import { getSlotInLocationStack } from '@opentrons/step-generation'
 
+import { getMissingTips } from '../utils/getMissingTips'
 import styles from './tippickupcontainer.module.css'
-import { getMissingTips } from './utils/getMissingTips'
 
 import type { TipType } from '@opentrons/components'
 import type { LabwareEntity, RobotState } from '@opentrons/step-generation'
