+++ app/src/organisms/Desktop/ProtocolVisualization/TipPickupContainer/tippickupcontainer.module.css
