+++ app/src/organisms/Desktop/ProtocolVisualization/WellTooltip/index.tsx
@@ -9,8 +9,8 @@ import { swatchColors } from '@opentrons/step-generation'
 
 import { getModalPortalEl } from '/app/App/portal'
 
-import { formatPercentage } from './utils/formatPercentage'
-import { formatVolume } from './utils/formatVolume'
+import { formatPercentage } from '../utils/formatPercentage'
+import { formatVolume } from '../utils/formatVolume'
 import styles from './welltooltip.module.css'
 
 import type { MouseEvent, ReactNode } from 'react'
